requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==40.0.1
defusedxml==0.7.1
dj-database-url==1.3.0
dj-rest-auth==3.0.0
dj-static==0.0.6
Django==4.1.7
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
install==1.3.5
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
